TritonInferenceServer/clients/requirements.txt
Brotli==1.1.0
certifi==2024.2.2
cmake==3.28.3
filelock==3.13.1
fsspec==2024.2.0
gevent==24.2.1
geventhttpclient==2.0.11
greenlet==3.0.3
Jinja2==3.1.3
lit==17.0.6
MarkupSafe==2.1.5
monai==1.3.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
pillow==10.2.0
python-rapidjson==1.14
six==1.16.0
sympy==1.12
torch==2.0.1
triton==2.0.0
tritonclient==2.42.0
typing_extensions==4.9.0
urllib3==2.2.1
zope.event==5.0
zope.interface==6.2
